+++ packages/ui/src/components/dialog/index.ts
@@ -1,5 +1,6 @@
 export { default as ArchDialog } from "./ArchDialog.vue";
 export { default as ArchDialogContent } from "./ArchDialogContent.vue";
+export type { ArchDialogContentSize } from "./ArchDialogContent.vue";
 export { default as ArchDialogDescription } from "./ArchDialogDescription.vue";
 export { default as ArchDialogFooter } from "./ArchDialogFooter.vue";
 export { default as ArchDialogHeader } from "./ArchDialogHeader.vue";