+++ packages/ui/src/components/dropdown/ArchDropdownItem.vue
@@ -16,7 +16,7 @@ const props = withDefaults(
   }
 );
 
-const dropdown = inject(dropdownContextKey);
+const dropdown = inject(dropdownContextKey, null);
 
 function selectItem() {
   if (!props.disabled) {