+++ packages/ui/src/components/toast/ArchToast.vue
@@ -11,6 +11,14 @@ const { dismiss } = useToast();
 function close() {
   dismiss(props.toast.id);
 }
+
+function runAction(action: NonNullable<ArchToastItem["actions"]>[number]) {
+  action.onClick();
+
+  if (action.dismiss !== false) {
+    close();
+  }
+}
 </script>
 
 <template>
@@ -25,6 +33,17 @@ function close() {
       </div>
       <div v-if="props.toast.description" class="arch-toast__description">
         {{ props.toast.description }}
+      </div>
+      <div v-if="props.toast.actions?.length" class="arch-toast__actions">
+        <button
+          v-for="(action, index) in props.toast.actions"
+          :key="`${props.toast.id}-${index}`"
+          type="button"
+          class="arch-toast__action"
+          @click="runAction(action)"
+        >
+          {{ action.label }}
+        </button>
       </div>
     </div>
     <button class="arch-toast__close" type="button" aria-label="Close" @click="close">