+++ packages/ui/src/components/virtual-scroller/index.ts
@@ -1,1 +1,2 @@
 export { default as ArchVirtualScroller } from "./ArchVirtualScroller.vue";
+export type { ArchVirtualScrollerProps } from "./ArchVirtualScroller.vue";